+++ tests/testthat/test-fc_wrappers.R
@@ -1,6 +1,8 @@
 context("wrappers")
 
-skip_if(Sys.getenv("FC_PRIVATE_KEY") == "")
+testthat::skip_on_cran()
+testthat::skip_if(Sys.getenv("FC_PRIVATE_KEY") == "")
+testthat::skip_if(!fc_ping())
 
 
 test_that("fc_get_user() works", {
@@ -33,18 +35,9 @@ test_that("fc_get_data_range() works", {
   data <- fc_get_data_range(
       station_id = stations[[1]]$station_name,
       data_group = "raw",
-      from = as.integer(as.POSIXct(Sys.time() - 60*60*24)),
+      from = as.integer(as.POSIXct(Sys.time() - 60 * 60 * 24)),
       to =  as.integer(as.POSIXct(Sys.time())))
   expect_is(data, "list")
   expect_true(names(data)[[1]] == "sensors")
 })
 
-test_that("fc_get_data_range() handles 204s", {
-
-  data <- fc_get_data_range(
-    station_id = stations[[1]]$station_name,
-    data_group = "raw",
-    from = as.integer(as.POSIXct(stations[[1]]$db$min_date) - 100),
-    to =  as.integer(as.POSIXct(stations[[1]]$db$min_date) - 10))
-  expect_equal(data, NULL)
-})