+++ NEWS.md
@@ -1,3 +1,10 @@
-# rfieldclimate 0.1.0.9000
+# rfieldclimate 0.1.1
+
+* Maintenance release
+* fc_request gained timeout argument
+* new function fc_ping to test api ping
+* Fail gracefully if API has errors or is not available
+
+# rfieldclimate 0.1.0
 
 * Initial release.